commit view staff profile funtion

+++ Backend/WCPMS/src/main/java/com/sep490/wcpms/repository/CustomerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
     // Spring Data JPA sẽ tự động tạo câu query để tìm Customer dựa trên ID của đối tượng Account liên kết
-    Optional<Customer> findByAccount_Id(Long accountId);
+    Optional<Customer> findByAccount_Id(Integer accountId);
 }
