+++ Backend/WCPMS/src/main/java/com/sep490/wcpms/service/ProfileService.java
@@ -4,6 +4,6 @@ import com.sep490.wcpms.dto.ProfileResponseDTO;
 import com.sep490.wcpms.dto.ProfileUpdateRequestDTO;
 
 public interface ProfileService {
-    ProfileResponseDTO getProfileByAccountId(Long accountId);
-    ProfileResponseDTO updateProfile(Long accountId, ProfileUpdateRequestDTO updateRequestDTO);
+    ProfileResponseDTO getProfileByAccountId(Integer accountId);
+    ProfileResponseDTO updateProfile(Integer accountId, ProfileUpdateRequestDTO updateRequestDTO);
 }
